Separate report query assembly
